Add participation service
