+++ src/endpoints/index.ts
@@ -1,6 +1,7 @@
 import Accounts from "./accounts";
-import Study from "./studies";
-
+import Study from "./study";
+import Participation from "./participation";
 import Auth from "./auth";
+import Studies from "./studies";
 
-export { Auth, Accounts, Study as Studies };
+export { Auth, Accounts, Study, Studies, Participation };
