+++ src/endpoints/study/researchers.ts
